Felix/damage visuals

+++ game/trees/scripts/Trees/Twee.gd
@@ -24,6 +24,7 @@ var life_time_seconds := 0.0
 #const TIME_TO_GROW = 5.0
 
 var is_large := false
+var is_growing := false
 
 func _ready():
 	get_stats_from_resource(tree_stat)
@@ -48,6 +49,46 @@ func _process(delta: float) -> void:
 			upgrade_tree()
 			#tree_data.update()
 
+
+
+#const FLASH_DECAY_RATE = 50.0
+const SHAKE_DECAY_RATE = 35.0
+const FLASH_DURATION = 0.1 # In seconds
+var flash_time = 0.0
+var flash_amount = 0.0
+var shake_amount = 0.0
+var is_dehydrated = false
+
+## Handle animating shader parameters relating to damage
+func _update_shader(delta: float) -> void:
+	if died: ## Death handles its own stuff.
+		return
+	
+	if flash_time > 0:
+		flash_amount = 1.0
+		flash_time -= delta
+		flash_time = max(flash_time, 0.0)
+	else:
+		flash_amount = 0.0
+	
+	if (!sprite):
+		return
+	shake_amount = move_toward(shake_amount, 0.0, delta * SHAKE_DECAY_RATE)
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("flash_amount", flash_amount)
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("shake_amount", shake_amount)
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("alpha", modulate.a)
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("pos", pos)
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("dehydrated", is_dehydrated)
+	#print(is_dehydrated)
+
+	# UV OFFSET FOR TRUNK DIFFERS BY LOCATION ON SHEET (Short and tall)
+	# IF MORE SPRITES ARE ADDED BELOW THE SHEET, BEWARE, MUST TWEAK VALUES!
+	if is_large:
+		(sprite.get_material() as ShaderMaterial).set_shader_parameter("uv_y_offset", 0.1)
+	else:
+		(sprite.get_material() as ShaderMaterial).set_shader_parameter("uv_y_offset", 0.62)
+
+
 ## NOTHING to SMALL
 func play_grow_small_animation():
 	animation_player.play("grow_small")
@@ -61,6 +102,7 @@ func _on_animation_player_animation_finished(anim_name: StringName) -> void:
 		animation_player.play("small")
 	if (anim_name == "grow_large"):
 		animation_player.play("large")
+		is_growing = false
 	if (anim_name == "die"):
 		animation_player.play("stump")
 
@@ -87,10 +129,22 @@ func initialize(p: Vector2i, f: int):
 	forest = f
 	init_pos(p)
 
+const GREEN_TREE_DIE = preload("res://trees/scenes/death/GreenTreeDie.tscn")
 
 func die():
 	died = true
 	#TreeManager.remove_tree(pos)
+	flash_amount = 1.0
+	(sprite.get_material() as ShaderMaterial).set_shader_parameter("flash_amount", flash_amount)
+	
+	await get_tree().create_timer(FLASH_DURATION).timeout
+	
+	if is_large and !is_growing: # Temp fix: Prevent small trees from spawning big tree die vfx
+		var death_vfx = GREEN_TREE_DIE.instantiate()
+		get_parent().add_child(death_vfx)
+		death_vfx.global_position = global_position
+		#queue_free()
+	
 	queue_free()
 	
 	#animation_player.play("die")
@@ -98,7 +152,12 @@ func die():
 		#func(animation_name):
 			#queue_free()
 	#)
-	
+
+const WATER_DAMAGE_DELAY = 3.0
+var water_damage_time := 0.0
+
+const DEHYDRATION_DAMAGE = 2
+
 ## update local storage and use water for maintainence
 ## returns the right amount of res to system
 func update(delta: float) -> Vector3:
@@ -108,22 +167,28 @@ func update(delta: float) -> Vector3:
 	storage = min(storage + gain.y, max_water)
 	
 	# take maint from storage
-	if (storage >= maint):
-		storage -= maint
-	else:
-		var f: Forest = TreeManager.get_forest(forest)
-		if (!f.get_water(maint - storage)):
-			# if game doesn't have enough water either
-			hp -= 2 * delta
-			print("DYING")
-			print(forest)
-		else:
-			# game has enough water
-			storage = 0
-	if (hp <= 0):
-		TreeManager.remove_tree(pos)
-		return Vector3()
+	#if (storage >= maint):
+		#storage -= maint
+		#is_dehydrated = false
+		#water_damage_time = 0.0
+	#else:
+		#var f: Forest = TreeManager.get_forest(forest)
+		#if (!f.get_water(maint - storage)):
+			## if game doesn't have enough water either
+			#is_dehydrated = true
+			#while (water_damage_time > WATER_DAMAGE_DELAY):
+				#take_damage(2)
+				#water_damage_time -= WATER_DAMAGE_DELAY
+			#water_damage_time += delta
+		#else:
+			## game has enough water
+			#is_dehydrated = false
+			#water_damage_time = 0.0
+			#storage = 0
+	#if (hp <= 0):
+		#die()
 	var g = Vector3(gain.x, storage - prev, gain.z)
+	#_update_shader(delta)
 	return g
 
 func update_maint():
@@ -142,11 +207,15 @@ func take_damage(damage: int) -> bool:
 	
 	hp -= damage
 	
-	if (hp <= 0):
+	if (hp <= 0 and TreeManager.get_tree_map()[pos]):
 		TreeManager.remove_tree(pos)
 		return true
 	#else:
 		#animation_player.play("hurt")
+	
+	# VISUAL: SET AMOUNTS FOR FLASH & SHAKE
+	flash_time = FLASH_DURATION
+	shake_amount = 20.0
 	
 	return false
 
@@ -154,4 +223,5 @@ func take_damage(damage: int) -> bool:
 func upgrade_tree() -> void:
 	is_large = true
 	animation_player.play("grow_large")
+	is_growing = true
 	get_upgraded_stats_from_resource(tree_stat)
